requirements.txt
aiohttp~=3.8.1
APScheduler~=3.10.4
python-dotenv~=1.0.0
deepl~=1.14.0
emoji<2.0.0
emoji_translate~=0.1.1
facebook-sdk~=3.1.0
facebook_business~=14.0.0
fastapi~=0.95.1
google-analytics-data~=0.13.2
google-cloud-vision~=3.1.1
gunicorn~=20.1.0
kneed~=0.8.1
spicy
#lightgbm~=3.3.2
matplotlib~=3.5.3
mlxtend~=0.22.0
nltk~=3.7
numpy~=1.24.3
oauth2client~=4.1.3
openai~=0.27.6
pandas~=2.0.1
Pillow~=9.5.0
plotly~=5.10.0
protobuf~=3.20.0
pydantic~=1.10.2
pymongo~=4.3.3
PyYAML~=6.0
regex~=2023.5.4
requests~=2.29.0
scikit-learn~=1.2.2
scipy~=1.10.1
seaborn~=0.12.0
sentence_transformers~=2.2.2
setuptools~=65.3.0
torch~=1.12.1
transformers~=4.28.1
urllib3~=1.26.12
xgboost~=1.6.2
python-dotenv~=1.0.0
keybert
keyphrase-vectorizers~=0.0.10
loguru~=0.7.0
pytz~=2023.3
uvicorn~=0.22.0
google~=3.0.0
dnspython~=2.3.0
google-api-python-client~=2.86.0
langdetect~=1.0.9
googlemaps~=4.10.0
google-cloud-aiplatform~=1.26.1
google-auth-oauthlib~=1.0.0
google-analytics-data